+++ src/honeyhive/_generated/services/Datapoints_service.py
@@ -16,11 +16,7 @@ def getDatapoints(
 
     base_path = api_config.base_path
     path = f"/v1/datapoints"
-    headers = {
-        "Content-Type": "application/json",
-        "Accept": "application/json",
-        "Authorization": f"Bearer { api_config.get_access_token() }",
-    }
+    headers = api_config.get_default_headers()
     query_params: Dict[str, Any] = {
         "datapoint_ids": datapoint_ids,
         "dataset_name": dataset_name,
@@ -58,11 +54,7 @@ def createDatapoint(
 
     base_path = api_config.base_path
     path = f"/v1/datapoints"
-    headers = {
-        "Content-Type": "application/json",
-        "Accept": "application/json",
-        "Authorization": f"Bearer { api_config.get_access_token() }",
-    }
+    headers = api_config.get_default_headers()
     query_params: Dict[str, Any] = {}
 
     query_params = {
@@ -102,11 +94,7 @@ def batchCreateDatapoints(
 
     base_path = api_config.base_path
     path = f"/v1/datapoints/batch"
-    headers = {
-        "Content-Type": "application/json",
-        "Accept": "application/json",
-        "Authorization": f"Bearer { api_config.get_access_token() }",
-    }
+    headers = api_config.get_default_headers()
     query_params: Dict[str, Any] = {}
 
     query_params = {
@@ -144,11 +132,7 @@ def getDatapoint(
 
     base_path = api_config.base_path
     path = f"/v1/datapoints/{id}"
-    headers = {
-        "Content-Type": "application/json",
-        "Accept": "application/json",
-        "Authorization": f"Bearer { api_config.get_access_token() }",
-    }
+    headers = api_config.get_default_headers()
     query_params: Dict[str, Any] = {}
 
     query_params = {
@@ -184,11 +168,7 @@ def updateDatapoint(
 
     base_path = api_config.base_path
     path = f"/v1/datapoints/{id}"
-    headers = {
-        "Content-Type": "application/json",
-        "Accept": "application/json",
-        "Authorization": f"Bearer { api_config.get_access_token() }",
-    }
+    headers = api_config.get_default_headers()
     query_params: Dict[str, Any] = {}
 
     query_params = {
@@ -226,11 +206,7 @@ def deleteDatapoint(
 
     base_path = api_config.base_path
     path = f"/v1/datapoints/{id}"
-    headers = {
-        "Content-Type": "application/json",
-        "Accept": "application/json",
-        "Authorization": f"Bearer { api_config.get_access_token() }",
-    }
+    headers = api_config.get_default_headers()
     query_params: Dict[str, Any] = {}
 
     query_params = {