+++ src/honeyhive/_generated/services/async_Datasets_service.py
@@ -16,11 +16,7 @@ async def getDatasets(
 
     base_path = api_config.base_path
     path = f"/v1/datasets"
-    headers = {
-        "Content-Type": "application/json",
-        "Accept": "application/json",
-        "Authorization": f"Bearer { api_config.get_access_token() }",
-    }
+    headers = api_config.get_default_headers()
     query_params: Dict[str, Any] = {"dataset_id": dataset_id, "name": name}
 
     query_params = {
@@ -55,11 +51,7 @@ async def createDataset(
 
     base_path = api_config.base_path
     path = f"/v1/datasets"
-    headers = {
-        "Content-Type": "application/json",
-        "Accept": "application/json",
-        "Authorization": f"Bearer { api_config.get_access_token() }",
-    }
+    headers = api_config.get_default_headers()
     query_params: Dict[str, Any] = {}
 
     query_params = {
@@ -97,11 +89,7 @@ async def updateDataset(
 
     base_path = api_config.base_path
     path = f"/v1/datasets"
-    headers = {
-        "Content-Type": "application/json",
-        "Accept": "application/json",
-        "Authorization": f"Bearer { api_config.get_access_token() }",
-    }
+    headers = api_config.get_default_headers()
     query_params: Dict[str, Any] = {}
 
     query_params = {
@@ -139,11 +127,7 @@ async def deleteDataset(
 
     base_path = api_config.base_path
     path = f"/v1/datasets"
-    headers = {
-        "Content-Type": "application/json",
-        "Accept": "application/json",
-        "Authorization": f"Bearer { api_config.get_access_token() }",
-    }
+    headers = api_config.get_default_headers()
     query_params: Dict[str, Any] = {"dataset_id": dataset_id}
 
     query_params = {
@@ -183,11 +167,7 @@ async def addDatapoints(
 
     base_path = api_config.base_path
     path = f"/v1/datasets/{dataset_id}/datapoints"
-    headers = {
-        "Content-Type": "application/json",
-        "Accept": "application/json",
-        "Authorization": f"Bearer { api_config.get_access_token() }",
-    }
+    headers = api_config.get_default_headers()
     query_params: Dict[str, Any] = {}
 
     query_params = {
@@ -228,11 +208,7 @@ async def removeDatapoint(
 
     base_path = api_config.base_path
     path = f"/v1/datasets/{dataset_id}/datapoints/{datapoint_id}"
-    headers = {
-        "Content-Type": "application/json",
-        "Accept": "application/json",
-        "Authorization": f"Bearer { api_config.get_access_token() }",
-    }
+    headers = api_config.get_default_headers()
     query_params: Dict[str, Any] = {}
 
     query_params = {