+++ src/honeyhive/_generated/services/async_Projects_service.py
@@ -13,11 +13,7 @@ async def getProjects(
 
     base_path = api_config.base_path
     path = f"/v1/projects"
-    headers = {
-        "Content-Type": "application/json",
-        "Accept": "application/json",
-        "Authorization": f"Bearer { api_config.get_access_token() }",
-    }
+    headers = api_config.get_default_headers()
     query_params: Dict[str, Any] = {"name": name}
 
     query_params = {
@@ -52,11 +48,7 @@ async def createProject(
 
     base_path = api_config.base_path
     path = f"/v1/projects"
-    headers = {
-        "Content-Type": "application/json",
-        "Accept": "application/json",
-        "Authorization": f"Bearer { api_config.get_access_token() }",
-    }
+    headers = api_config.get_default_headers()
     query_params: Dict[str, Any] = {}
 
     query_params = {
@@ -92,11 +84,7 @@ async def updateProject(
 
     base_path = api_config.base_path
     path = f"/v1/projects"
-    headers = {
-        "Content-Type": "application/json",
-        "Accept": "application/json",
-        "Authorization": f"Bearer { api_config.get_access_token() }",
-    }
+    headers = api_config.get_default_headers()
     query_params: Dict[str, Any] = {}
 
     query_params = {
@@ -132,11 +120,7 @@ async def deleteProject(
 
     base_path = api_config.base_path
     path = f"/v1/projects"
-    headers = {
-        "Content-Type": "application/json",
-        "Accept": "application/json",
-        "Authorization": f"Bearer { api_config.get_access_token() }",
-    }
+    headers = api_config.get_default_headers()
     query_params: Dict[str, Any] = {"name": name}
 
     query_params = {